+++ src/ai/chat.rs
@@ -4,19 +4,25 @@ use std::{
 };
 
 use console::{strip_ansi_codes, style, Term};
+use futures::StreamExt;
 use indicatif::{ProgressBar, TermLike};
+use rig::{message::Message, streaming::StreamingChoice};
 use rustyline::DefaultEditor;
 use termimad::MadSkin;
 use textwrap::wrap;
 
 use crate::{
     ai::utils::{on_the_fly_change_model, NO_MODELS_FOUND_MSG},
-    client::{ChatMessage, ChatRole},
     config::Config,
     utils::console::get_spinner_style,
 };
 
-pub fn chat(term: &Term, cfg: &Config, mut initial_message: Option<String>, select_model: bool) {
+pub async fn chat(
+    term: &Term,
+    cfg: &Config,
+    mut initial_message: Option<String>,
+    select_model: bool,
+) {
     let mut provider = cfg
         .active_provider()
         .unwrap_or_else(|| {
@@ -27,7 +33,7 @@ pub fn chat(term: &Term, cfg: &Config, mut initial_message: Option<String>, sele
 
     if select_model {
         println!();
-        if let Some(p) = on_the_fly_change_model(&mut cfg.clone(), Some(provider.model())) {
+        if let Some(p) = on_the_fly_change_model(&mut cfg.clone(), Some(provider.model())).await {
             provider = p;
         } else {
             println!("{}", style(NO_MODELS_FOUND_MSG).red());
@@ -60,7 +66,8 @@ pub fn chat(term: &Term, cfg: &Config, mut initial_message: Option<String>, sele
 
         if select_model {
             select_model = false;
-            if let Some(p) = on_the_fly_change_model(&mut cfg.clone(), Some(provider.model())) {
+            if let Some(p) = on_the_fly_change_model(&mut cfg.clone(), Some(provider.model())).await
+            {
                 provider = p;
                 println!();
             } else {
@@ -83,7 +90,9 @@ pub fn chat(term: &Term, cfg: &Config, mut initial_message: Option<String>, sele
             }
         }
 
-        if input.trim().eq("/model") {
+        input = input.trim().to_string();
+
+        if input.eq("/model") {
             select_model = true;
             println!();
             continue;
@@ -91,12 +100,12 @@ pub fn chat(term: &Term, cfg: &Config, mut initial_message: Option<String>, sele
 
         println!();
 
-        if exit_words.contains(&input.trim().to_lowercase()) {
+        if exit_words.contains(&input.to_lowercase()) {
             println!("{ai}\nGoodbye! 👋\n");
             std::process::exit(0);
         }
 
-        if input.trim() == "clear" {
+        if input.eq("clear") {
             term.clear_screen().expect("Failed to clear screen");
             messages.clear();
             println!("{ai}\nWhat can I help with?\n");
@@ -106,27 +115,41 @@ pub fn chat(term: &Term, cfg: &Config, mut initial_message: Option<String>, sele
         spinner.enable_steady_tick(Duration::from_millis(100));
         spinner.set_message(format!("{ai}"));
 
-        messages.push(ChatMessage {
-            role: ChatRole::User,
-            content: input.trim().into(),
-        });
-
         if cfg.streaming() {
             let mut response = String::new();
-            let content_iter = provider.chat_stream(&messages);
+            let mut final_response = String::new();
+
+            let mut stream = provider.chat_stream(&input, messages.clone()).await;
+
             let mut line_count = 0;
 
             term.hide_cursor().expect("Failed to hide cursor");
 
-            for (i, content) in content_iter.enumerate() {
+            let mut clear = true;
+            while let Some(chunk) = stream.next().await {
+                let Ok(choice) = chunk else {
+                    continue;
+                };
+
+                let content = match choice {
+                    StreamingChoice::Message(content) => content,
+                    StreamingChoice::ToolCall(name, bob, params) => {
+                        let tool = format!("Tool: {name} - {bob} - {params}");
+                        println!("{tool}");
+                        continue;
+                    }
+                };
+
                 let term_width = term.width() as usize;
 
-                if i == 0 {
+                if clear {
+                    clear = false;
                     spinner.finish_and_clear();
                     println!("{ai}");
                 }
 
                 response.push_str(&content);
+                final_response.push_str(&content);
 
                 let text = format!("{}", skin.text(&response, Some(term_width)));
 
@@ -144,15 +167,21 @@ pub fn chat(term: &Term, cfg: &Config, mut initial_message: Option<String>, sele
                 }
             }
 
+            messages.push(Message::user(&input));
+            messages.push(Message::assistant(final_response));
+
             term.show_cursor().expect("Failed to show cursor");
             term.flush().expect("Failed to flush terminal");
         } else {
-            let response = provider.chat(&messages);
+            let response = provider.chat(&input, messages.clone()).await;
 
             spinner.finish_and_clear();
             println!("{ai}");
             skin.print_text(&response);
             println!();
+
+            messages.push(Message::user(&input));
+            messages.push(Message::assistant(response));
         }
     }
 }