+++ src/client/mod.rs
@@ -3,6 +3,4 @@ mod constants;
 mod models;
 mod traits;
 
-pub use models::shared::{ChatMessage, ChatRole, ContentIterator};
-
 pub use ai_client::AIClient as Client;