+++ src/client/traits.rs
@@ -1,8 +1,3 @@
-pub trait Choice {}
-pub trait ContentTrait {
-    fn extract_content(&self) -> Option<String>;
-}
-
 pub trait ModelTrait {
     fn extract_models(&self) -> Result<Vec<(String, String)>, String>;
 }