+++ src/main.rs
@@ -2,13 +2,15 @@ mod ai;
 mod args;
 mod client;
 mod config;
+// mod mcp;
 mod program;
 mod provider;
 mod utils;
 
 use program::Program;
 
-fn main() {
+#[tokio::main]
+async fn main() {
     ctrlc::set_handler(move || {
         let term = console::Term::stdout();
         term.flush().expect("Failed to flush terminal");
@@ -17,5 +19,11 @@ fn main() {
     })
     .expect("Failed to set Ctrl-C handler");
 
-    Program::run();
+    // let command = "deno";
+    // let args = vec!["-A", "/home/lr/Development/mcp-playground/server/main.ts", "--stdio"];
+    // let client = mcp::McpClient::new(command, args).await;
+    // client.tools().await;
+    // client.cancel().await;
+
+    Program::run().await;
 }