+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.0
+
+### Added
+
+- Added async runtime with `tokio` to enable support for Model Context Protocol (MCP) in future releases.
+
+### Fixed
+
+- Fixed an issue that caused assistant responses to not be added to the message history in chat sessions.
+
 ## 0.3.0
 
 ### Added